presidium-analyzer/requirements.txt
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz
spacy
grpcio
futures
